bersih2

+++ app/Http/Controllers/CommentController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers;
 use App\Models\Comment;
 use App\Models\CommentLike;
 use App\Models\Post;
-use Illuminate\Http\Request;
 
 class CommentController extends Controller
 {
@@ -14,6 +13,7 @@ class CommentController extends Controller
         $fields = request()->validate([
             'content' => "required|min:3|max:250"
         ]);
+        
         $fields["user_id"] = auth()->user()->id;
         $fields["post_id"] = $post->id;
         $fields["comment_id"] = null;
@@ -23,30 +23,24 @@ class CommentController extends Controller
     }
 
     // like/unlike comment
-    public function like(Comment $comment)
+    public function like(Comment $comment, CommentLike $commentLike)
     {
-        $like = CommentLike::where('user_id', auth()->user()->id)
-            ->where('comment_id', $comment->id)->first();
+        $userId = auth()->user()->id;
+        $like = $comment->getLikeByUser($userId);
 
-        if ($like) {
-            $like->delete();
-            return response()->json([
-                'liked' => false,
-                'likeCount' => CommentLike::where('comment_id', $comment->id)->count()
-            ]);
-        }
-        CommentLike::create(['comment_id' => $comment->id, 'user_id' => auth()->user()->id]);
+        if ($like) $like->delete();
+        else $commentLike->create(['comment_id' => $comment->id, 'user_id' => auth()->user()->id]);
+
         return response()->json([
-            'liked' => true,
-            'likeCount' => CommentLike::where('comment_id', $comment->id)->count()
+            'liked' => $like ? false : true,
+            'likeCount' => $comment->likes->count()
         ]);
     }
 
     public function destroy(Comment $comment)
     {
-        if ($comment->user_id !== auth()->user()->id) {
-            abort(403, "Unauthorized");
-        }         
+        if ($comment->user_id !== auth()->user()->id) abort(403, "Unauthorized");
+
         $comment->delete();
         return back();
     }
